requirements.txt
astor==0.8.1
attrs==21.4.0
bandit==1.7.4
certifi==2021.10.8
darglint==1.8.1
docutils==0.18.1
elastic-transport==8.1.1
elasticsearch==8.1.2
eradicate==2.0.0
flake8==3.9.2
flake8-bandit==2.1.2
flake8-broken-line==0.4.0
flake8-bugbear==21.11.29
flake8-commas==2.1.0
flake8-comprehensions==3.8.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-eradicate==1.2.0
flake8-isort==4.1.1
flake8-polyfill==1.0.2
flake8-quotes==3.3.1
flake8-rst-docstrings==0.2.5
flake8-string-format==0.3.0
gitdb==4.0.9
GitPython==3.1.27
isort==5.10.1
mccabe==0.6.1
pbr==5.8.1
pep8-naming==0.12.1
psycopg2-binary==2.9
pycodestyle==2.7.0
pydantic==1.9.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.11.2
python-dotenv==0.20.0
PyYAML==6.0
restructuredtext-lint==1.4.0
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==3.5.0
testfixtures==6.18.5
tqdm==4.63.1
typing_extensions==4.1.1
urllib3==1.26.9
wemake-python-styleguide==0.16.0
